% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_criteria.R
\name{assess_criterion_a}
\alias{assess_criterion_a}
\title{Evaluate Criterion A (Population Reduction)}
\usage{
assess_criterion_a(trend_res)
}
\arguments{
\item{trend_res}{Result from calculate_trend()}
}
\value{
Suggested category (character)
}
\description{
Evaluate Criterion A (Population Reduction)
}
