% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_criteria.R
\name{assess_criterion_d2}
\alias{assess_criterion_d2}
\title{Evaluate Criterion D2 (Restricted AOO/Locations)}
\usage{
assess_criterion_d2(aoo_km2, locations)
}
\arguments{
\item{aoo_km2}{Numeric AOO}

\item{locations}{Numeric Location count}
}
\value{
"VU" or "LC"
}
\description{
Evaluate Criterion D2 (Restricted AOO/Locations)
}
