% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_criteria.R
\name{assess_species}
\alias{assess_species}
\title{Run Full IUCN Assessment for a Single Species (Harmonised Pipeline)}
\usage{
assess_species(sp, recent_start, recent_end, comp_start, comp_end)
}
\arguments{
\item{sp}{Character. Species scientific name.}

\item{recent_start}{Numeric. Start year for recent data.}

\item{recent_end}{Numeric. End year for recent data.}

\item{comp_start}{Numeric. Start year for comparison data.}

\item{comp_end}{Numeric. End year for comparison data.}
}
\description{
Run Full IUCN Assessment for a Single Species (Harmonised Pipeline)
}
