% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_assess.R
\name{batch_assess}
\alias{batch_assess}
\title{Batch IUCN Assessment (Harmonised Temporal Scale)}
\usage{
batch_assess(
  species_list,
  recent_start = 2010,
  recent_end = 2026,
  comp_start = 2000,
  comp_end = 2016
)
}
\arguments{
\item{species_list}{Vector of scientific names.}

\item{recent_start}{Numeric. Start year for recent data.}

\item{recent_end}{Numeric. End year for recent data.}

\item{comp_start}{Numeric. Start year for comparison data.}

\item{comp_end}{Numeric. End year for comparison data.}
}
\description{
Batch IUCN Assessment (Harmonised Temporal Scale)
}
