% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{clean_dates}
\alias{clean_dates}
\title{Clean and Parse NDOP Dates}
\usage{
clean_dates(df)
}
\arguments{
\item{df}{Data frame with DATUM_OD column}
}
\value{
Data frame with added 'year' column
}
\description{
Clean and Parse NDOP Dates
}
