% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mature_terms.R
\name{get_mature_terms}
\alias{get_mature_terms}
\title{Get List of Mature Life Stages}
\usage{
get_mature_terms()
}
\value{
A character vector of terms.
}
\description{
Defines the specific terms from NDOP that represent reproductive capability
for both animals and plants.
}
