% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_assessment.R
\name{summarize_assessment}
\alias{summarize_assessment}
\title{Summarize Assessment with Strict IUCN Cascade Logic}
\usage{
summarize_assessment(
  species,
  eoo,
  aoo,
  trend,
  locations,
  pop_metrics,
  evaluate_pop = TRUE,
  year_last = NA,
  n_records = NA,
  a_criteria = c("A2")
)
}
\arguments{
\item{species}{Character string. Species name.}

\item{eoo}{List containing \code{area_km2}.}

\item{aoo}{List containing \code{area_km2}.}

\item{trend}{List containing \code{percent_change}.}

\item{locations}{Numeric. Number of locations.}

\item{pop_metrics}{List containing \code{decline_rate}, \code{fluct_ratio}, \code{total_mature}, and \code{max_subpop}.}

\item{evaluate_pop}{Logical. If FALSE, ignores Criteria C, D1, and pop-dependent subcriteria.}

\item{year_last}{Numeric. The most recent year observed in the dataset.}

\item{n_records}{Numeric. Total number of records in the dataset.}
}
\value{
A list containing \code{result} (display dataframe) and \code{details} (boolean flags).
}
\description{
Evaluates Criteria A-E, plus Pre-Assessment for RE (Extinct) and DD (Data Deficient).
}
