% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_assess.R
\name{batch_assess}
\alias{batch_assess}
\title{Batch IUCN Assessment (Harmonized Temporal Scale)}
\usage{
batch_assess(species_list, recent_years = 10)
}
\arguments{
\item{species_list}{Vector of scientific names.}

\item{recent_years}{Numeric. Window for current status (default 10).}
}
\description{
Batch IUCN Assessment (Harmonized Temporal Scale)
}
