% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aoo.R
\name{calculate_aoo}
\alias{calculate_aoo}
\title{Calculate Area of Occupancy (AOO)}
\usage{
calculate_aoo(occurrence_data, grid_size = 2000)
}
\arguments{
\item{occurrence_data}{An sf object (points).}

\item{grid_size}{Numeric. Grid cell size in meters. Default 2000 (2km).}
}
\value{
A list containing the AOO area in km2 and the grid geometry.
}
\description{
Calculate Area of Occupancy (AOO)
}
