% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_aoo.R
\name{calculate_aoo}
\alias{calculate_aoo}
\title{Calculate AOO with Temporal Filter}
\usage{
calculate_aoo(occurrence_data, grid_size = 2000, year_start = NULL)
}
\arguments{
\item{occurrence_data}{sf object}

\item{grid_size}{Numeric (default 2000 for IUCN)}

\item{year_start}{Numeric. Only records from this year onwards.}
}
\description{
Calculate AOO with Temporal Filter
}
