% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_eoo.R
\name{calculate_eoo}
\alias{calculate_eoo}
\title{Calculate EOO with Temporal Filter}
\usage{
calculate_eoo(occurrence_data, year_start = NULL)
}
\arguments{
\item{occurrence_data}{sf object}

\item{year_start}{Numeric. Only records from this year onwards.}
}
\description{
Calculate EOO with Temporal Filter
}
