% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_eoo.R
\name{calculate_eoo}
\alias{calculate_eoo}
\title{Calculate Extent of Occurrence (EOO)}
\usage{
calculate_eoo(occurrence_data)
}
\arguments{
\item{occurrence_data}{An sf object (points).}
}
\value{
A list containing the EOO area in km2 and the polygon geometry.
}
\description{
Calculate Extent of Occurrence (EOO)
}
