% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iucn.R
\name{plot_iucn}
\alias{plot_iucn}
\title{Plot IUCN Assessment Map}
\usage{
plot_iucn(species_name, occ_data = NULL, window = 10)
}
\arguments{
\item{species_name}{Character.}

\item{occ_data}{An existing sf object (optional). If NULL, it will download.}

\item{window}{Numeric.}
}
\description{
Plot IUCN Assessment Map
}
